refactor: :recycle: Remove meteor functions from `reloadRoomAfterLogin.ts`
Removed Meteor.Startup, Tracker.autorun, Meteor.user and direct usage of the router. Also implemented ref logic for the currentUsername verification and tried to fix the roomType if.
