+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -7,6 +7,7 @@ import { useForceLogout } from './hooks/useForceLogout';
 import { useGoogleTagManager } from './hooks/useGoogleTagManager';
 import { useMessageLinkClicks } from './hooks/useMessageLinkClicks';
 import { useOTRMessaging } from './hooks/useOTRMessaging';
+import { useReloadRoomAfterLogin } from './hooks/useReloadRoomAfterLogin';
 import { useSettingsOnLoadSiteUrl } from './hooks/useSettingsOnLoadSiteUrl';
 import { useStoreCookiesOnLogin } from './hooks/useStoreCookiesOnLogin';
 import { useUpdateVideoConfUser } from './hooks/useUpdateVideoConfUser';
@@ -69,6 +70,7 @@ const AppLayout = () => {
 	useAutoupdate();
 	useForceLogout();
 	useNotificationUserCalendar();
+	useReloadRoomAfterLogin();
 
 	const layout = useSyncExternalStore(appLayout.subscribe, appLayout.getSnapshot);
